Split TestContext implementation
Move the live output scope, scoped context setter, and synchronized string builder into focused partial-class files without changing behavior or API.

Copilot-Session: 071532a3-358a-44ef-9ccd-a694c48f539b
